Adding apple module

+++ MedievalWar/Engine/Components/VelocityComponent.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 struct VelocityComponent: Component {
-    var maxSpeed: Double
-    var x: Double
-    var y: Double
+    var maxSpeed: Float
+    var x: Float
+    var y: Float
 }
